+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 * Add "Duplicate" integration action to organization integration view. Refs UIORGS-441.
 * *BREAKING* Add number generator for vendor code including settings page. Refs UIORGS-336, UIORGS-337.
 * Enhancement help text on Settings > Organizations > Number generator options. Refs UIORGS-453.
+* React v19: refactor away from react-test-renderer. Refs UIORGS-433.
+* React v19: refactor away from default props for functional components. Refs UIORGS-434.
 
 ## [5.2.0](https://github.com/folio-org/ui-organizations/tree/v5.2.0) (2024-10-31)
 [Full Changelog](https://github.com/folio-org/ui-organizations/compare/v5.1.1...v5.2.0)